(Fix) Staff Dashboard Backup Delete Function

+++ app/Http/Controllers/Staff/BackupController.php
@@ -111,16 +111,20 @@ class BackupController extends Controller
      * Deletes A Backup.
      *
      * @param \Illuminate\Http\Request $request
-     * @param $file_name
      */
-    public function delete(Request $request, $file_name)
+    public function delete(Request $request)
     {
         $disk = Storage::disk($request->input('disk'));
+        $file_name = $request->input('file_name');
+        $adapter = $disk->getDriver()->getAdapter();
 
-        if ($disk->exists($file_name)) {
-            $disk->delete($file_name);
-
-            return 'success';
+        if ($adapter instanceof Local) {
+            if ($disk->exists($file_name)) {
+                $disk->delete($file_name);
+                return 'success';
+            } else {
+                return abort(404, trans('backup.backup_doesnt_exist'));
+            }
         }
 
         return abort(404, trans('backup.backup_doesnt_exist'));
